pyproject.toml
[project]
name = "ai-usage"
version = "0.1.0"
description = "Track AI subscription usage across Claude, ChatGPT, and GitHub Copilot"
readme = "README.md"
authors = [
    { name = "Ariel Sandor", email = "39200214+arielsandor@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "textual>=3.1.0",
    "typer>=0.15.0",
    "httpx>=0.28.0",
    "keyring>=25.6.0",
    "pyyaml>=6.0",
    "rich>=14.0.0",
    "pydantic>=2.11.0",
]

[project.scripts]
ai-usage = "ai_usage.main:app"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
